+++ test/liveValidatorTests.ts
@@ -5,7 +5,6 @@ import * as assert from "assert";
 import * as os from "os";
 import * as path from "path";
 import * as lodash from "lodash";
-import { ResponsesObject } from "yasway";
 import { LiveValidator } from "../lib/liveValidation/liveValidator";
 import { OperationSearcher } from "../lib/liveValidation/operationSearcher";
 import * as Constants from "../lib/util/constants";
@@ -523,7 +522,12 @@ describe("Live Validator", () => {
 
       // Operations to match is StorageAccounts_CheckNameAvailability with provider "Hello.World"
       // [non cached provider]
-      validationInfo = validator.parseValidationRequest(nonCachedProviderUrl, "PoSt", "randomId", "");
+      validationInfo = validator.parseValidationRequest(
+        nonCachedProviderUrl,
+        "PoSt",
+        "randomId",
+        ""
+      );
       result = validator.operationSearcher.getPotentialOperations(validationInfo);
       operations = result.matches;
       reason = result.reason;
@@ -577,7 +581,7 @@ describe("Live Validator", () => {
       const operations = microsoftTest.get("2016-01-01")?.get("post")!;
 
       for (const operation of operations) {
-        const responses = operation.responses as ResponsesObject;
+        const responses = operation.responses;
         assert.strictEqual(responses.default, undefined);
       }
     });
@@ -711,7 +715,9 @@ describe("Live Validator", () => {
     it(`should pass response header tests`, async () => {
       const options = {
         directory: `./test/liveValidation/swaggers/`,
-        swaggerPathsPattern: ["specification/apimanagement/resource-manager/Microsoft.ApiManagement/**/*.json"],
+        swaggerPathsPattern: [
+          "specification/apimanagement/resource-manager/Microsoft.ApiManagement/**/*.json",
+        ],
       };
       const validator = new LiveValidator(options);
       await validator.initialize();
@@ -836,16 +842,17 @@ describe("Live Validator", () => {
           const payload = require(`${__dirname}/liveValidation/payloads/xmsSecretAndPOST/xmsSecretButGet_${payloadVersion}.json`);
           const result = await liveValidator.validateLiveRequestResponse(payload);
           assert.equal(result.responseValidationResult.isSuccessful, false);
-					const errors = result.responseValidationResult.errors;
-					for (const error of errors) {
-						assert.equal(
-							(error.schemaPath.indexOf("x-ms-secret") !== -1 && error.code === "SECRET_PROPERTY") ||
-								(error.schemaPath.indexOf("x-ms-mutability") !== -1 &&
-									error.code === "WRITEONLY_PROPERTY_NOT_ALLOWED_IN_RESPONSE"),
-							true
-						);
-					}
-				}
+          const errors = result.responseValidationResult.errors;
+          for (const error of errors) {
+            assert.equal(
+              (error.schemaPath.indexOf("x-ms-secret") !== -1 &&
+                error.code === "SECRET_PROPERTY") ||
+                (error.schemaPath.indexOf("x-ms-mutability") !== -1 &&
+                  error.code === "WRITEONLY_PROPERTY_NOT_ALLOWED_IN_RESPONSE"),
+              true
+            );
+          }
+        }
       });
     });